Category delete implemented

+++ scripts/modules/categories/categoriesController.js
@@ -1,7 +1,7 @@
 (function(){
 	'use strict';
 	var categoriesController	=	angular.module('categoriesController',[])
-		.controller('categoriesController',function($scope,categoriesFact,$state,categoryService){
+		.controller('categoriesController',function($scope,categoriesFact,$state,Notification,categoryService){
 			categoriesFact.viewCategories()
 				.then(function(responseData){
 					categoryService.setData(responseData.data)
@@ -10,12 +10,18 @@
 				})
 
 			$scope.delete = function(id){
-				categoriesFact.deleteCategory(id)
-					.then(function(success){
-						console.log(success);
-					},function(error){
-						console.log(error)
-					})
+				var conf = confirm('Are You sure to Delete');
+				if(conf == true) {
+					categoriesFact.deleteCategory(id)
+						.then(function(success){
+							Notification.success('Success notification');
+				    		categoryService.spliceData(id,1);
+						},function(error){
+							Notification.warning({message: 'Errorr', title: 'Error Occured'});
+						})
+					}else{
+						return false;
+					}
 			}
 		
 		})
@@ -23,14 +29,14 @@
 				$scope.save = function(category){
 				var $parent_id = (typeof(category.parent_id) != 'undefined') ? category.parent_id : '';
 				var $data = {name: category.name,transaction_type:category.transaction_type,parent_id : $parent_id};
-				console.log(category);
+				//console.log(category);
 				categoriesFact.addCategories($data)
 					.then(function(success){
 						console.log(success.data)
-						$data.id = success.data.id;
-						$data.parent_id = success.data.parent_id == "" ? 0 : 1;
-						categoryService.pushData($data);
-						$state.go('index.categories');
+						// $data.id = success.data.id;
+						// $data.parent_id = success.data.parent_id == "" ? 0 : 1;
+						// categoryService.pushData($data);
+						// $state.go('index.categories');
 					},function(error){
 
 					})
